+++ src/main/java/pt/weldtrack/weld/track_api/controller/EquipamentoController.java
@@ -1,5 +1,6 @@
 package pt.weldtrack.weld.track_api.controller;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import pt.weldtrack.weld.track_api.model.EquipamentoEntity;
@@ -10,6 +11,8 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/equipamentos")
 public class EquipamentoController {
+
+    private final String CHAVE = "weldtrack123";
 
     private final EquipamentoService service;
 
@@ -34,7 +37,13 @@ public class EquipamentoController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id) {
+    public ResponseEntity<String> delete(
+            @PathVariable Long id,
+            @RequestHeader(value = "X-Admin-Key", required = false) String chaveRecebida) {
+        if (chaveRecebida == null || !chaveRecebida.equals(CHAVE)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Erro: nao tem autorizaçãoo para criar registos");
+        }
         service.delete(id);
+        return ResponseEntity.ok("Registo " + id + " apagado com sucesso.");
     }
 }
